Fix naming and move code to the dedicated package
