+++ bencode/deserialize/deserialize_test.go
@@ -1,4 +1,4 @@
-package bencode
+package deserialize
 
 import (
 	"strings"
